Made the timer more obvious.
Changes include the following:
When the time is more than 30, the timer text colour will alternate between green and yellow.
When the time is less than 30, the timer text will alternate between red and yellow.
Whenever the timer is reset, the timer will flash between red and yellow rapidly, to try to catch the player's attention.