Tried to make controls more obvious
I've made the tutorial at the bridge slightly transparent, so its less obstructive to the player.
I've attached a script to the tutorial thing that makes it boop up and down, though we now have two things that are trying to catch the player's attention, might be a bit overwhelming? But the script can be disabled if deemed unneeded.
The same tutorial image has been added to the pause menu, so that the player can be sure to find it if they press the pause button.