More zombies, and added some logic to the re-enabling of zombies
Tested on my laptop, 50 zombies seems good, though the laptop is subpar, and maybe not a good enough test?
From all zombies on the map, a set number is chosen, where we then randomly pick this amount from the pool of zombies and enable them.
When a zombie is killed, the GameObject is destroyed after 20seconds, a new zombie is picked from the pool and enabled right away.