Fixed Danger Zones and Zombies walk on Lava
There were some danger zones marked as not on trigger, these are now marked.
There were an issue with the lava from the volcano marked as convex and trigger, causing weird death zones by the powerplant. Made the cylinder not convex and made invisible death zones.
Made the capsule collider of the zombies go into the ground, which should make it touch the plane of the lava, killing the zombie, after driving around alot, I have yet to spot a zombie that could walk on Lava.