Added an Endgame Button and made the time Expo friendly.
The Endgame button has been added to the pause screen, pushing the end game button will take the player to the score screen to submit the score.

The timer has been revamped, it will start at 180 seconds, then it will at 1/10th of the time that was in previous builds to the timer. It has been tested that it is impossible to play infinite. Expected play time for a player that plays well is 5-6min.

The changes can be made to the manager which is located in:
Structures -> Bus Stops -> TimeManager, where the multiplier is the amount of time added. It was 0.1 in the past, and is now 0.01. Initialtime is for the start time, currently set to 180 seconds.